ai-recommendation-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic>=2.11.0,<3
pydantic-settings>=2.10.1,<3
PyJWT==2.10.1
redis>=5.0.0,<6
networkx==3.5
httpx>=0.27.0,<1
sqlalchemy[asyncio]>=2.0.0,<2.1
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
pgvector>=0.3.0
numpy>=1.26.0
sentence-transformers>=2.7.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
